+++ packages/node/src/configure/SubqueryProject.spec.ts
@@ -1,6 +1,7 @@
 // Copyright 2020-2023 SubQuery Pte Ltd authors & contributors
 // SPDX-License-Identifier: GPL-3.0
 
+import fs from 'fs';
 import path from 'path';
 
 import {
@@ -11,7 +12,27 @@ import {
   makeTempDir,
   ReaderFactory,
 } from '@subql/common';
-import { Reader } from '@subql/types-core';
+import {
+  SubstrateCustomDataSourceImpl,
+  isCustomDs,
+} from '@subql/common-substrate';
+import {
+  loadDataSourceScript,
+  saveFile,
+  SubqlProjectDs,
+  updateDataSourcesEntry,
+  updateDataSourcesV1_0_0,
+  updateProcessor,
+} from '@subql/node-core';
+import {
+  SubstrateCustomDatasource,
+  SubstrateRuntimeDatasource,
+} from '@subql/types';
+import {
+  BaseCustomDataSource,
+  BaseDataSource,
+  Reader,
+} from '@subql/types-core';
 import { SubqueryProject } from './SubqueryProject';
 
 // eslint-disable-next-line jest/no-export
@@ -21,6 +42,16 @@ export async function getProjectRoot(reader: Reader): Promise<string> {
     return makeTempDir();
   }
   throw new Error('Un-known reader type');
+}
+
+// Inspect asset path and content
+function inspectAsset(inspectAssetPath: string) {
+  const { dir, ext } = path.parse(inspectAssetPath);
+  // should have no extension
+  expect(ext).toBe('');
+  const assetContent = fs.readFileSync(inspectAssetPath, 'utf8');
+  // And when we load it, it should resolve correct abi
+  expect(assetContent).toContain('transferFrom');
 }
 
 describe('SubqueryProject', () => {
@@ -145,9 +176,77 @@ describe('SubqueryProject', () => {
       },
     );
     expect(project.templates[0].name).toBeDefined();
-    // Expect file path to be ipfs too
-    expect((project.templates[0] as any).assets.get('erc20').file).toBe(
-      'ipfs://QmYoHL3BvEW6nH1zYZqnziUHjajadu5ErJHavHS2zXkZhv',
+    // Expect asset to be fetched
+    const inspectAssetPath = (project.templates[0] as any).assets.get(
+      'erc20',
+    ).file;
+    inspectAsset(inspectAssetPath);
+  }, 50000);
+});
+
+describe('load asset with updateDataSourcesV1_0_0', () => {
+  const customDsImpl: SubstrateCustomDataSourceImpl[] = [
+    {
+      kind: 'substrate/FrontierEvm',
+      assets: new Map([
+        [
+          'erc20',
+          {
+            file: 'ipfs://QmYoHL3BvEW6nH1zYZqnziUHjajadu5ErJHavHS2zXkZhv',
+          },
+        ],
+      ]),
+      mapping: {
+        file: 'ipfs://QmP4Hrfydh4zswkZYeTnnZQFhTGo3LkCfHz4jdkbP8ZA8P',
+        handlers: [
+          {
+            filter: {
+              topics: [
+                'Transfer(address indexed from,address indexed to,uint256 value)',
+              ],
+            },
+            handler: 'handleEvmEvent',
+            kind: 'substrate/FrontierEvmEvent',
+          },
+          {
+            filter: {
+              function: 'approve(address to,uint256 value)',
+            },
+            handler: 'handleEvmCall',
+            kind: 'substrate/FrontierEvmCall',
+          },
+        ],
+      },
+      processor: {
+        file: 'ipfs://QmeHHtqRFSJQwv8pr6oUDsAkNPNAPSXXPoKiab8NKJHkiH',
+        options: {
+          abi: 'erc20',
+          address: '0x6bd193ee6d2104f14f94e2ca6efefae561a4334b',
+        },
+      },
+      startBlock: 752073,
+      validate: jest.fn(),
+    },
+  ];
+  let root: string;
+  let reader: Reader;
+  beforeEach(async () => {
+    reader = await ReaderFactory.create(
+      'ipfs://QmRoosV27325uAeepKqaTEPFKjC3nk4rrKmZJSd7QXYKZQ',
+      { ipfs: IPFS_NODE_ENDPOINT },
     );
+    root = await makeTempDir();
+  });
+  afterEach(() => {
+    fs.rmSync(root, { recursive: true, force: true });
+  });
+
+  it('After fix, it could load asset correctly', async () => {
+    const ds = await updateDataSourcesV1_0_0<
+      SubstrateRuntimeDatasource,
+      SubstrateCustomDatasource<any, any>
+    >(customDsImpl, reader, root, isCustomDs);
+    const inspectAssetPath = (ds[0] as any).assets.get('erc20').file;
+    inspectAsset(inspectAssetPath);
   }, 50000);
 });